requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.1
sentence-transformers==2.7.0
scikit-learn==1.3.2
reportlab==4.0.7
PyPDF2==3.0.1
openai-whisper==20231117
python-dotenv==1.0.0
huggingface_hub==0.23.0
